评论，回复接口开发完成

+++ black-hole-app/src/main/java/com/free/badmood/blackhole/web/entity/Comment.java
@@ -37,7 +37,7 @@ public class Comment extends BaseEntity {
     /**
      * 评论用户id
      */
-    private Integer fromUserId;
+    private long fromUserId;
